# docs/openapi.yaml
openapi: 3.0.3
info:
  title: REGS — 線上自動評測系統 API
  version: "1.0.0"
  description: |
    CS3060701 期末專案。權限階層 Guest < User < Admin。
    標示 security 的端點需 `Authorization: Bearer <JWT>`(ES256 / EC P-256)。
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    TokenResponse:
      type: object
      properties:
        token: { type: string }
        token_type: { type: string, example: Bearer }
        user:
          type: object
          properties:
            id: { type: string }
            username: { type: string }
            role: { type: string, enum: [user, admin] }
    Problem:
      type: object
      properties:
        id: { type: string, example: "113final006" }
        title: { type: string }
        description: { type: string }
        cases:
          type: array
          items: { type: string }
    CaseResult:
      type: object
      properties:
        case_name: { type: string }
        status: { type: string, enum: [AC, WA, CE, RE, TLE, SE, IE] }
        exit_code: { type: integer }
        duration_ms: { type: integer, format: int64 }
    Submission:
      type: object
      properties:
        operatorId: { type: string }
        user_id: { type: string }
        problem_id: { type: string }
        status:
          type: string
          enum: [PENDING, RUNNING, AC, WA, CE, RE, TLE, SE, IE]
        message: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        case_results:
          type: array
          items: { $ref: "#/components/schemas/CaseResult" }
paths:
  /api/users/register:
    post:
      summary: 註冊新使用者(Guest)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, minLength: 4 }
      responses:
        "201": { description: created }
        "409": { description: username taken, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/users/login:
    post:
      summary: 登入並簽發 JWT(Guest)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/TokenResponse" } } } }
        "401": { description: invalid credentials }
  /api/users/logout:
    post:
      summary: 登出(User)
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: ok } }
  /api/users/me:
    get:
      summary: 取得目前登入者資料(User)
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: ok }, "401": { description: unauthorized } }
  /api/users/{user_id}/submissions:
    get:
      summary: 取得指定使用者的提交紀錄(Guest)
      security: []
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: ok } }
  /api/problems:
    get:
      summary: 題目列表(Guest)
      security: []
      responses: { "200": { description: ok } }
    put:
      summary: 建立或更新題目(Admin)。JSON metadata 或 multipart(欄位 archive 上傳題目壓縮檔)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
                title: { type: string }
                description: { type: string }
          multipart/form-data:
            schema:
              type: object
              properties:
                id: { type: string }
                title: { type: string }
                description: { type: string }
                archive: { type: string, format: binary }
      responses: { "200": { description: updated }, "201": { description: created }, "401": { description: unauthorized } }
  /api/problems/{problem_id}:
    get:
      summary: 單一題目描述(Guest)
      security: []
      parameters: [{ name: problem_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Problem" } } } }, "404": { description: not found } }
    delete:
      summary: 刪除題目(Admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: problem_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: deleted }, "401": { description: unauthorized } }
  /api/problems/{problem_id}/testcases:
    get:
      summary: 下載題目壓縮檔(Admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: problem_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: archive (binary) }, "404": { description: no archive } }
  /api/submissions:
    post:
      summary: 提交程式碼進行評測(User)。立即回 operatorId,背景評測
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [problem_id, file]
              properties:
                problem_id: { type: string }
                file: { type: string, format: binary, description: "zip/tar/tgz,不含前導資料夾" }
      responses:
        "202":
          description: accepted
          content: { application/json: { schema: { type: object, properties: { operatorId: { type: string }, status: { type: string } } } } }
        "401": { description: unauthorized }
    get:
      summary: 查詢個人提交紀錄列表(User)
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: ok } }
  /api/submissions/{operatorId}:
    get:
      summary: 取得單一提交評測結果(User;Admin 可查他人)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: operatorId, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Submission" } } } }
        "403": { description: not your submission }
        "404": { description: not found }
  /api/submissions/{operatorId}/source:
    get:
      summary: 下載提交的原始專案檔(User;Admin 可查他人)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: operatorId, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: archive (binary) }, "404": { description: missing } }
  /api/submissions/{operatorId}/logs/{stage}:
    get:
      summary: 多分段日誌查詢(User;Admin 可查他人)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: operatorId, in: path, required: true, schema: { type: string } }
        - { name: stage, in: path, required: true, schema: { type: string, enum: [configure, compile, output] } }
      responses:
        "200": { description: ok, content: { application/json: { schema: { type: object, properties: { stage: { type: string }, log: { type: string } } } } } }
        "400": { description: bad stage }
  /api/submissions/{operatorId}/rerun:
    post:
      summary: 重新執行某個評測 Job(User;Admin 可重跑他人)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: operatorId, in: path, required: true, schema: { type: string } }]
      responses: { "202": { description: re-queued } }
  /api/stats/problems/{problem_id}:
    get:
      summary: 題目統計(Guest)
      security: []
      parameters: [{ name: problem_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: ok } }
  /api/stats/users/{user_id}:
    get:
      summary: 使用者統計(Guest)
      security: []
      parameters: [{ name: user_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: ok } }